+++ assets/js/device_manager.js
@@ -50,7 +50,9 @@
     installed_devices: {},
 
     init: function() {
-      console.log('init device_manager.js');
+      if (this.debug) {
+        console.log('init device_manager.js');
+      }
       this.default_plugin = DeviceManagerDefaultPlugin;
       // call the init function on each plugin if it defines one
       Object.keys(this.plugins).forEach(function (key) {
@@ -76,15 +78,15 @@
       }
     },
     get_device: function(dev) {
-      if (this.debug) console.log("get_device", dev)
+      if (this.debug) { console.log("get_device", dev); }
       return this.devices[dev]
     },
     get_devices: function() {
-      if (this.debug) console.log("get_devices")
+      if (this.debug) { console.log("get_devices"); }
       return { devices: this.devices, current_device: this.devices.current_device }
     },
     set_devices: function(devs) {
-      if (this.debug) console.log("set_devices", devs)
+      if (this.debug) { console.log("set_devices", devs); }
       var installed_devices = this.installed_devices;
 
       this.set_devices_id(devs, installed_devices, "handsfree_input_id", "default")
@@ -94,7 +96,7 @@
       this.set_devices_id(devs, installed_devices, "video_input_id", "")
     },
     set_devices_id: function(devs, installed, type, the_default) {
-      if (this.debug) console.log("set_devices_id", devs, installed, type, the_default)
+      if (this.debug) { console.log("set_devices_id", devs, installed, type, the_default); }
       if (installed[devs[type]])
         this.devices[type] = devs[type]
       else
@@ -106,19 +108,20 @@
         console.log('has_headset_device devices', this.devices);
         console.log('has_headset_device installed devices', this.installed_devices);
       }
-      if(this.devices.headset_input_id && this.devices.headset_output_id)
+      if(this.devices.headset_input_id && this.devices.headset_output_id) {
         status = true;
-      if (this.debug) console.log("has_headset_device", status)
+      }
+      if (this.debug) { console.log("has_headset_device", status); }
       return status;
     },
     enumerateDevices: function() {
-      if (DeviceManager.debug) console.log('device manager enumerateDevices')
+      if (DeviceManager.debug) { console.log('device manager enumerateDevices'); }
       navigator.mediaDevices.enumerateDevices()
       .then(DeviceManager.saveDevices)
       .catch(DeviceManager.errorCallback)
     },
     saveDevices: function(deviceInfos) {
-      if (DeviceManager.debug) console.log('saveDevices deviceInfos', deviceInfos)
+      if (DeviceManager.debug) { console.log('saveDevices deviceInfos', deviceInfos); }
       DeviceManager.installed_devices = {}
       let devices = []
       for (var i = 0; i !== deviceInfos.length; ++i) {
@@ -131,7 +134,7 @@
         }
       }
       UcxUcc.installed_devices = devices
-      if (DeviceManager.debug) console.log('installed_devices', devices)
+      if (DeviceManager.debug) { console.log('installed_devices', devices); }
 
       setTimeout(() => {
         var event = new Event('device_manager_init');
@@ -139,24 +142,25 @@
       }, 100);
     },
     load_devices: function() {
-      if (DeviceManager.debug) console.log("load_devices()")
+      if (DeviceManager.debug) { console.log("load_devices()"); }
       navigator.mediaDevices.enumerateDevices()
       .then(DeviceManager.gotDevices)
       .catch(DeviceManager.errorCallback)
     },
     get_current_device: function() {
-      if (this.debug) console.log("get_current_device")
+      if (this.debug) { console.log("get_current_device"); }
       return this.devices.current_device;
     },
     setSinkId: function(audio_control, sinkId) {
-      if (DeviceManager.debug) console.log("setSinkId", audio_control, sinkId)
+      if (DeviceManager.debug) { console.log("setSinkId", audio_control, sinkId); }
       let element = audio_control[0]
       if (sinkId) {
         if (typeof element.sinkId != 'undefined') {
           element.setSinkId(sinkId)
           .then(function() {
-            if (DeviceManager.debug)
-              console.log('Success, audio output device attached', sinkId)
+            if (DeviceManager.debug) {
+              console.log('Success, audio output device attached', sinkId);
+            }
           })
           .catch(function(error) {
             var errorMessage = error;
@@ -166,11 +170,10 @@
             }
           });
         } else {
-          console.error("Browser does not support output device selection.")
+          console.error("Browser does not support output device selection.");
         }
       } else {
-        if (DeviceManager.debug)
-          console.log('Ignoring setSinkId for no sinkId', sinkId)
+        if (DeviceManager.debug) { console.log('Ignoring setSinkId for no sinkId', sinkId); }
       }
     },
     set_sink_id_headset_output_id_audio: function() {
@@ -192,7 +195,7 @@
       this.devices.current_device = this.devices.handsfree_input_id;
     },
     stop: function(audio_control) {
-      if (DeviceManager.debug) console.log("device manager stop")
+      if (DeviceManager.debug) { console.log("device manager stop"); }
       audio_control[0].pause()
       audio_control.attr('src','')
     },
@@ -230,7 +233,7 @@
       if (this.debug) console.log("volume_decrement")
     },
     call_on_hold: function(key, audio_ctrl) {
-      if (this.debug) { console.log('call_on_hold, key', key, 'audio_ctrl', audio_ctrl) }
+      if (this.debug) { console.log('call_on_hold, key', key, 'audio_ctrl', audio_ctrl); }
       if (audio_ctrl.attr('src')) {
         if (key == 0) {
           audio_ctrl[0].play()
@@ -261,11 +264,11 @@
     },
     start: function(args) {
       this.is_device_manager = true;
-      console.log('starting device_manager with args', args)
+      console.log('starting device_manager with args', args);
     },
     stop: function(args) {
       this.is_device_manager = false;
-      console.log('stopping device_manager with args', args)
+      console.log('stopping device_manager with args', args);
     },
     volume_up: function() {},
     volume_set: function(args) {
@@ -286,7 +289,7 @@
   document.dispatchEvent(new Event('DeviceManagerLoaded'));
 
   UccChat.on_connect(function(ucc_chat, socket) {
-    console.log('device_manager on_connect');
+    if (DeviceManager.debug) { console.log('device_manager on_connect'); }
     window.UccChat.DeviceManager = DeviceManager;
     setTimeout(function() {
       window.UccChat.DeviceManager.init();